fix: formatting

+++ librawstorio/src/poll_queue.cpp
@@ -227,7 +227,8 @@ rawstor::io::Event* Queue::recvmsg(
     std::function<void(size_t, int)>&& cb
 ) {
     TraceEvent trace_event = RAWSTOR_TRACE_EVENT(
-        '|', "fd = %d, niov = %u, flags = %u\n", fd, msg->msg_iovlen, flags
+        '|', "fd = %d, niov = %u, flags = %u\n", fd,
+        (unsigned int)msg->msg_iovlen, flags
     );
     Session& s = _get_session(fd);
 
@@ -335,7 +336,8 @@ rawstor::io::Event* Queue::sendmsg(
     std::function<void(size_t, int)>&& cb
 ) {
     TraceEvent trace_event = RAWSTOR_TRACE_EVENT(
-        '|', "fd = %d, niov = %u, flags = %u\n", fd, msg->msg_iovlen, flags
+        '|', "fd = %d, niov = %u, flags = %u\n", fd,
+        (unsigned int)msg->msg_iovlen, flags
     );
     Session& s = _get_session(fd);
